move ColumnCatalog from frontend crate to common crate
