+++ src/common/src/catalog/column.rs
@@ -12,9 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use itertools::Itertools;
-use risingwave_pb::plan_common::ColumnDesc as ProstColumnDesc;
+use std::borrow::Cow;
 
+use itertools::Itertools;
+use risingwave_pb::plan_common::{
+    ColumnCatalog as ProstColumnCatalog, ColumnDesc as ProstColumnDesc,
+};
+
+use super::row_id_column_desc;
 use crate::catalog::Field;
 use crate::error::ErrorCode;
 use crate::types::DataType;
@@ -219,6 +224,69 @@ impl From<&ColumnDesc> for ProstColumnDesc {
             name: c.name.clone(),
             field_descs: c.field_descs.iter().map(ColumnDesc::to_protobuf).collect(),
             type_name: c.type_name.clone(),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ColumnCatalog {
+    pub column_desc: ColumnDesc,
+    pub is_hidden: bool,
+}
+
+impl ColumnCatalog {
+    /// Get the column catalog's is hidden.
+    pub fn is_hidden(&self) -> bool {
+        self.is_hidden
+    }
+
+    /// Get a reference to the column desc's data type.
+    pub fn data_type(&self) -> &DataType {
+        &self.column_desc.data_type
+    }
+
+    /// Get the column desc's column id.
+    pub fn column_id(&self) -> ColumnId {
+        self.column_desc.column_id
+    }
+
+    /// Get a reference to the column desc's name.
+    pub fn name(&self) -> &str {
+        self.column_desc.name.as_ref()
+    }
+
+    /// Convert column catalog to proto
+    pub fn to_protobuf(&self) -> ProstColumnCatalog {
+        ProstColumnCatalog {
+            column_desc: Some(self.column_desc.to_protobuf()),
+            is_hidden: self.is_hidden,
+        }
+    }
+
+    /// Creates a row ID column (for implicit primary key).
+    pub fn row_id_column() -> Self {
+        Self {
+            column_desc: row_id_column_desc(),
+            is_hidden: true,
+        }
+    }
+}
+
+impl From<ProstColumnCatalog> for ColumnCatalog {
+    fn from(prost: ProstColumnCatalog) -> Self {
+        Self {
+            column_desc: prost.column_desc.unwrap().into(),
+            is_hidden: prost.is_hidden,
+        }
+    }
+}
+
+impl ColumnCatalog {
+    pub fn name_with_hidden(&self) -> Cow<'_, str> {
+        if self.is_hidden {
+            Cow::Owned(format!("{}(hidden)", self.column_desc.name))
+        } else {
+            Cow::Borrowed(&self.column_desc.name)
         }
     }
 }
