+++ src/frontend/src/catalog/mod.rs
@@ -18,13 +18,11 @@
 //! structs. It is accessed via [`catalog_service::CatalogReader`] and
 //! [`catalog_service::CatalogWriter`], which is held by [`crate::session::FrontendEnv`].
 
-use risingwave_common::catalog::{ColumnDesc, PG_CATALOG_SCHEMA_NAME};
+use risingwave_common::catalog::{is_row_id_column_name, PG_CATALOG_SCHEMA_NAME, ROWID_PREFIX};
 use risingwave_common::error::{ErrorCode, Result, RwError};
-use risingwave_common::types::DataType;
 use thiserror::Error;
 pub(crate) mod catalog_service;
 
-pub(crate) mod column_catalog;
 pub(crate) mod database_catalog;
 pub(crate) mod function_catalog;
 pub(crate) mod index_catalog;
@@ -72,35 +70,6 @@ pub fn check_schema_writable(schema: &str) -> Result<()> {
         )).into())
     } else {
         Ok(())
-    }
-}
-
-const ROWID_PREFIX: &str = "_row_id";
-
-pub fn row_id_column_name() -> String {
-    ROWID_PREFIX.to_string()
-}
-
-pub fn is_row_id_column_name(name: &str) -> bool {
-    name.starts_with(ROWID_PREFIX)
-}
-
-/// The column ID preserved for the row ID column.
-pub const ROW_ID_COLUMN_ID: ColumnId = ColumnId::new(0);
-
-/// The column ID offset for user-defined columns.
-///
-/// All IDs of user-defined columns must be greater or equal to this value.
-pub const USER_COLUMN_ID_OFFSET: i32 = ROW_ID_COLUMN_ID.next().get_id();
-
-/// Creates a row ID column (for implicit primary key). It'll always have the ID `0` for now.
-pub fn row_id_column_desc() -> ColumnDesc {
-    ColumnDesc {
-        data_type: DataType::Int64,
-        column_id: ROW_ID_COLUMN_ID,
-        name: row_id_column_name(),
-        field_descs: vec![],
-        type_name: "".to_string(),
     }
 }
