+++ src/frontend/src/catalog/system_catalog/mod.rs
@@ -22,7 +22,7 @@ use std::sync::{Arc, LazyLock};
 use async_trait::async_trait;
 use paste::paste;
 use risingwave_common::catalog::{
-    ColumnDesc, SysCatalogReader, TableDesc, TableId, DEFAULT_SUPER_USER_ID,
+    ColumnCatalog, ColumnDesc, SysCatalogReader, TableDesc, TableId, DEFAULT_SUPER_USER_ID,
     INFORMATION_SCHEMA_SCHEMA_NAME, PG_CATALOG_SCHEMA_NAME, RW_CATALOG_SCHEMA_NAME,
 };
 use risingwave_common::error::Result;
@@ -30,7 +30,6 @@ use risingwave_common::row::OwnedRow;
 use risingwave_common::types::DataType;
 
 use crate::catalog::catalog_service::CatalogReader;
-use crate::catalog::column_catalog::ColumnCatalog;
 use crate::catalog::system_catalog::information_schema::*;
 use crate::catalog::system_catalog::pg_catalog::*;
 use crate::catalog::system_catalog::rw_catalog::*;
